Final : 99% done.
The project is 99% finished and will be subject to further in-depth study.
A few critical security measures are lacking and will be individually implemented in the next projects, such as password encryption and token protection.
Note: Implementing AJAX to the framework is a task to look for in the near future.